Add files via upload
In our System Analysis and Design project, we developed a comprehensive model for an AI-powered sewing system. The project included creating multiple diagrams to depict different aspects of the system. We designed a use case diagram to outline the interactions between users and the system, an Entity-Relationship Diagram (ERD) to map out data structure, and a context diagram to show the overall system and its external interactions. We also developed a logical DFD (Data Flow Diagram) to represent the flow of data within the system, a child DFD to provide a more detailed view of individual processes, and a physical DFD to illustrate the physical implementation of the system. This approach provided a complete understanding of the system’s requirements, data flow, and functional design.