Add files via upload
In this networking project, we developed an ideal conceptual model for connecting the main network and subnetworks of various departments within the college. The project focused on designing a structured, efficient network architecture that facilitates seamless communication and data sharing between departments, while optimizing performance and security across the entire network.